pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sensebase"
version = "0.1.0"
description = "Extract knowledge from GitLab repositories for AI agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jason Baldwin"}
]
dependencies = [
    "python-gitlab>=4.0.0",
    "PyGithub>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
vectors = [
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "tiktoken>=0.5.0",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
]
full = [
    "sensebase[vectors]",
    "sensebase[api]",
    "tree-sitter>=0.21.0",
    "whoosh>=2.7.4",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
]

[project.scripts]
sensebase = "src.main:main"
sb-search = "src.query.search:main"
sb-semantic = "src.query.embeddings:main"
sb-api = "src.api.server:main"
sb-llm = "src.extractors.llm_extractor:main"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
